// frontend/packages/api-client/src/generated/schema.d.ts
export interface paths {
// ...
    "/auth/login": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** Password login */
        post: operations["mizan_apps_identity_api_auth_login"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
// ...
export interface operations {
    mizan_apps_audit_api_list_audit_events: {
// ...
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["Page_AuditEventOut_"];
                };
            };
        };
    };
// ...
}